Add UpdateTags method.  Update swagger docs

+++ api/timeline.go
@@ -51,6 +51,7 @@ func (service Service) GetAllTimelines(rw http.ResponseWriter, req *http.Request
 // @Tags timeline
 // @Accept  json
 // @Produce  json
+// @Param endpoint body api.Timeline true "The timeline to add"
 // @Success 200 {object} api.SystemResponse
 // @Failure 500 {object} api.ErrorResponse
 // @Router /timelines [put]
@@ -97,9 +98,10 @@ func (service Service) AddTimeline(rw http.ResponseWriter, req *http.Request) {
 // @Tags timeline
 // @Accept  json
 // @Produce  json
+// @Param id path string true "The timeline id to get"
 // @Success 200 {object} api.SystemResponse
 // @Failure 500 {object} api.ErrorResponse
-// @Router /timelines [put]
+// @Router /timelines/{id} [get]
 func (service Service) GetTimeline(rw http.ResponseWriter, req *http.Request) {
 
 	//	Get the id from the url (if it's blank, return an error)
@@ -134,13 +136,21 @@ func (service Service) GetTimeline(rw http.ResponseWriter, req *http.Request) {
 // @Tags timeline
 // @Accept  json
 // @Produce  json
+// @Param id path string true "The timeline id to delete"
 // @Success 200 {object} api.SystemResponse
+// @Failure 400 {object} api.ErrorResponse
 // @Failure 500 {object} api.ErrorResponse
-// @Router /timelines [delete]
+// @Router /timelines/{id} [delete]
 func (service Service) DeleteTimeline(rw http.ResponseWriter, req *http.Request) {
 
 	//	Get the id from the url (if it's blank, return an error)
 	timelineId := chi.URLParam(req, "id")
+
+	if timelineId == "" {
+		err := fmt.Errorf("requires an id of a timeline to delete")
+		sendErrorResponse(rw, err, http.StatusBadRequest)
+		return
+	}
 
 	//	Add a timeline
 	err := service.DB.DeleteTimeline(req.Context(), timelineId)
@@ -153,6 +163,58 @@ func (service Service) DeleteTimeline(rw http.ResponseWriter, req *http.Request)
 	//	Construct our response
 	response := SystemResponse{
 		Message: fmt.Sprintf("Timeline deleted: %v", timelineId),
+		Data:    timelineId,
+	}
+
+	//	Serialize to JSON & return the response:
+	rw.Header().Set("Content-Type", "application/json; charset=utf-8")
+	json.NewEncoder(rw).Encode(response)
+
+}
+
+// UpdateTags godoc
+// @Summary Updates tags for a timeline
+// @Description Updates tags for a timeline
+// @Tags timeline
+// @Accept  json
+// @Produce  json
+// @Param id path string true "The timeline id to update tags for"
+// @Param endpoint body api.UpdateTagsRequest true "The tags to set for the timeline"
+// @Success 200 {object} api.SystemResponse
+// @Failure 400 {object} api.ErrorResponse
+// @Failure 500 {object} api.ErrorResponse
+// @Router /timelines/{id} [post]
+func (service Service) UpdateTags(rw http.ResponseWriter, req *http.Request) {
+
+	//	Get the id from the url (if it's blank, return an error)
+	timelineId := chi.URLParam(req, "id")
+
+	if timelineId == "" {
+		err := fmt.Errorf("requires an id of a file to update tags for")
+		sendErrorResponse(rw, err, http.StatusBadRequest)
+		return
+	}
+
+	//	Parse the body to get the tags
+	request := UpdateTagsRequest{}
+	err := json.NewDecoder(req.Body).Decode(&request)
+	if err != nil {
+		err = fmt.Errorf("problem decoding tag update request: %v", err)
+		sendErrorResponse(rw, err, http.StatusBadRequest)
+		return
+	}
+
+	//	Update the timeline
+	err = service.DB.UpdateTags(req.Context(), timelineId, request.Tags)
+	if err != nil {
+		err = fmt.Errorf("error updating tags: %v", err)
+		sendErrorResponse(rw, err, http.StatusInternalServerError)
+		return
+	}
+
+	//	Construct our response
+	response := SystemResponse{
+		Message: "Timeline updated",
 		Data:    timelineId,
 	}
